requirements.txt
asgiref==3.5.2
Django==4.0.6
django-model-utils==4.2.0
django-tinymce==3.4.0
Pillow==9.2.0
sqlparse==0.4.2
tzdata==2022.1
